refactor: 길찾기 관련 타입 types로 분리
