+++ app/lib/getRoute.ts
@@ -1,16 +1,5 @@
 import axios from '@/app/lib/instance';
-
-type Coordinates = [number, number];
-
-type Direction = {
-  path: Coordinates[];
-  summary: {
-    distance: number; // 거리 (m)
-    duration: number; // 소요 시간 (ms)
-    taxiFare: number; // 택시 요금
-    tollFare: number; // 통행료
-  };
-};
+import { Direction } from '../types/directions';
 
 type RouteResponse = {
   route: {
